perf(nav): Phase 2 — bundle trimming, Stream lazy-load, #248 storm, RQ tuning
Bundle / deps (#636, #639):
- Remove dead deps (no imports): npm, swr, axios, cmdk, react-day-picker, tweetnacl(+util)
- prettier, dotenv -> devDependencies; bcryptjs seed -> bcrypt; pin @next/bundle-analyzer to ^15
- next.config: images.formats avif+webp, broaden serverExternalPackages, drop react-day-picker
  from transpilePackages
- Footer: react-icons/fa6 + FaXTwitter (the FA5 FaTwitter bird is deprecated post-rebrand)

Lazy-load the Stream SDK (#639, #248):
- StreamProvider is now an SDK-free shell that dynamic()-loads StreamProviderImpl (ssr:false);
  the SDK + its 2 stylesheets live only in the lazy chunk
- new lib/stream/disconnect.ts (SDK-free) owns the shared client refs + disconnectStreamClients,
  so Navbar and the other dashboards stop statically pulling the SDK
- #248: split the ChatSidebar mega-effect (initial fetch keyed on client+scope; listeners keyed
  on client) + in-flight guard so channel clicks no longer refire the queryChannels storm;
  defer connect to idle; HomeTabs read the video client at click-time and lazy-import lib/meeting
- recharts: lazy-load the admin payouts chart (extracted to PayoutsChart, dynamic ssr:false)

React Query (#636):
- mutations: { retry: 0 } (no auto-retry on payment/booking submits) + streaming dehydration default

Needs manual testing: chat open/switch/pagination, unread badge on home, meeting join from both
home tabs, logout/disconnect from every surface. Deferred: server caching (2.1), RSC conversion,
framer-motion LazyMotion.
