+++ package.json
@@ -98,30 +98,23 @@
     "@types/redlock": "^4.0.8",
     "@upstash/ratelimit": "^2.0.8",
     "@upstash/redis": "^1.36.1",
-    "axios": "^1.13.3",
     "bad-words": "^4.0.0",
     "bcrypt": "^6.0.0",
-    "bcryptjs": "^3.0.3",
     "better-auth": "^1.6.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cmdk": "^1.1.1",
     "date-fns": "^4.1.0",
     "date-fns-tz": "^3.2.0",
-    "dotenv": "^17.2.3",
     "framer-motion": "^11.15.0",
     "iso-639-1": "^3.1.5",
     "libsodium-wrappers": "^0.8.2",
     "lucide-react": "^0.563.0",
     "next": "^15.3.8",
     "next-nprogress-bar": "^2.4.7",
-    "npm": "^11.8.0",
     "pg": "^8.17.2",
-    "prettier": "^3.8.1",
     "razorpay": "^2.9.6",
     "react": "^18.3.1",
     "react-cookie-consent": "^10.0.1",
-    "react-day-picker": "^9.13.0",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.71.1",
     "react-icons": "^5.5.0",
@@ -132,18 +125,15 @@
     "stream-chat": "^9.30.1",
     "stream-chat-react": "^13.13.4",
     "stripe": "^20.2.0",
-    "swr": "^2.3.8",
     "tailwind-merge": "^3.4.0",
     "tailwindcss-animate": "^1.0.7",
-    "tweetnacl": "^1.0.3",
-    "tweetnacl-util": "^0.15.1",
     "use-debounce": "^10.1.0",
     "zod": "^3.25.67"
   },
   "devDependencies": {
     "@faker-js/faker": "^10.2.0",
     "@eslint/js": "^9.39.2",
-    "@next/bundle-analyzer": "^16.1.5",
+    "@next/bundle-analyzer": "^15",
     "fast-check": "^3.23.2",
     "@next/eslint-plugin-next": "^16.2.2",
     "@testing-library/jest-dom": "^6.9.1",
@@ -155,6 +145,7 @@
     "@types/react-dom": "^18.3.1",
     "autoprefixer": "^10.4.23",
     "chalk": "^5.6.2",
+    "dotenv": "^17.2.3",
     "dotenv-cli": "^11.0.0",
     "eslint": "^9.39.2",
     "eslint-plugin-jest": "^29.12.1",
@@ -166,6 +157,7 @@
     "jest": "^30.2.0",
     "jest-environment-jsdom": "^30.2.0",
     "postcss": "^8.5.6",
+    "prettier": "^3.8.1",
     "prisma": "7.7.0",
     "tailwindcss": "^3.4.17",
     "ts-node": "^10.9.2",
