+++ prisma/seedFiles/15a-create-organizations.ts
@@ -30,7 +30,7 @@
  */
 
 import { faker } from "@faker-js/faker";
-import bcrypt from "bcryptjs";
+import bcrypt from "bcrypt";
 import {
   Currency,
   FundingSource,
