+++ providers/StreamProvider.tsx
@@ -1,88 +1,21 @@
 "use client";
 
-import {
-  useCallback,
-  useEffect,
-  useState,
-  useRef,
-  createContext,
-  useContext,
-} from "react";
-import { StreamChat } from "stream-chat";
-import { Chat } from "stream-chat-react";
-import { StreamVideo, StreamVideoClient } from "@stream-io/video-react-sdk";
-import {
-  chatTokenProvider,
-  tokenProvider,
-} from "@/actions/stream/chat/stream.action";
-import { upsertUserToStream } from "@/actions/stream/chat/user.action";
-import { syncUserEventChannels } from "@/actions/stream/chat/event-channel.action";
-import { useUserData } from "@/hooks/useUserData";
-import { mapRoleToStream } from "@/lib/user";
-import { streamLogger } from "@/lib/stream-logger";
-import { clearAllStreamCaches } from "@/lib/stream-cache";
+import { createContext, useContext } from "react";
+import dynamic from "next/dynamic";
 
-// Client-side only: tracks which users have completed initial sync within this
-// browser tab's module lifecycle. Separate from the server-side Set in stream-cache.ts.
-const clientSyncCompletedUsers = new Set<string>();
-import StreamErrorBoundary from "@/components/stream/StreamErrorBoundary";
+// Re-export the logout helper from the SDK-free module so existing importers of
+// `disconnectStreamClients` from this path keep working unchanged. Callers that
+// ONLY need disconnect should import it straight from "@/lib/stream/disconnect"
+// to stay off the SDK bundle entirely (Navbar already does — see #248).
+export { disconnectStreamClients } from "@/lib/stream/disconnect";
 
-// Stream CSS co-located with the provider that renders Stream UI — the
-// video stylesheet was previously imported by both dashboard layouts.
-import "stream-chat-react/dist/css/v2/index.css";
-import "@stream-io/video-react-sdk/dist/css/styles.css";
+// ── Connection-state context ────────────────────────────────────────────────
+// Defined in this SDK-free shell (not the heavy impl) so consumers like
+// DebugDialog can import useStreamConnection without pulling the SDK, and so the
+// context identity is stable across the lazy boundary. The heavy impl imports
+// this same context and pushes the live value into it once connected.
 
-const apiKey = process.env.NEXT_PUBLIC_STREAM_API_KEY;
-
-// Module-level client instances to avoid re-creating on remount
-let globalChatClient: StreamChat | null = null;
-let globalVideoClient: StreamVideoClient | null = null;
-let currentUserId: string | null = null;
-
-/**
- * Disconnect all Stream clients and clear global state.
- * Call this before signing out to ensure clean disconnection.
- */
-export async function disconnectStreamClients(): Promise<void> {
-  const promises: Promise<void>[] = [];
-
-  if (globalChatClient) {
-    const client = globalChatClient;
-    promises.push(
-      client.disconnectUser().then(() => {
-        streamLogger.debug("Chat client disconnected on logout");
-      }),
-    );
-  }
-
-  if (globalVideoClient) {
-    const videoClient = globalVideoClient;
-    promises.push(
-      videoClient.disconnectUser().then(() => {
-        streamLogger.debug("Video client disconnected on logout");
-      }),
-    );
-  }
-
-  await Promise.all(promises);
-
-  // Nullify references only after disconnect completes to avoid
-  // another code path creating a new client while still disconnecting
-  globalChatClient = null;
-  globalVideoClient = null;
-  currentUserId = null;
-  clearAllStreamCaches();
-
-  // Clear sessionStorage sync flags so a re-login triggers a fresh sync
-  if (typeof sessionStorage !== "undefined") {
-    Object.keys(sessionStorage)
-      .filter((k) => k.startsWith("stream_sync_"))
-      .forEach((k) => sessionStorage.removeItem(k));
-  }
-}
-
-// Connection state context
-interface StreamConnectionState {
+export interface StreamConnectionState {
   chatConnected: boolean;
   videoConnected: boolean;
   isConnecting: boolean;
@@ -90,499 +23,80 @@ interface StreamConnectionState {
   retryConnection: () => void;
 }
 
-const StreamConnectionContext = createContext<StreamConnectionState | null>(
-  null,
-);
-
-export const useStreamConnection = () => {
-  const context = useContext(StreamConnectionContext);
-  if (!context) {
-    throw new Error("useStreamConnection must be used within StreamProvider");
-  }
-  return context;
+// Default (impl-not-yet-loaded) connection state. Returned by
+// useStreamConnection while the dynamic impl is still loading so consumers never
+// crash during the lazy-load window. Previously useStreamConnection threw when
+// used outside the provider; the dashboard never relied on that throw in prod,
+// and a safe default is required now that the provider is lazy.
+const DEFAULT_CONNECTION_STATE: StreamConnectionState = {
+  chatConnected: false,
+  videoConnected: false,
+  isConnecting: false,
+  error: null,
+  retryConnection: () => {},
 };
 
-interface StreamProviderProps {
+export const StreamConnectionContext = createContext<StreamConnectionState>(
+  DEFAULT_CONNECTION_STATE,
+);
+
+export const useStreamConnection = (): StreamConnectionState => {
+  return useContext(StreamConnectionContext);
+};
+
+export interface StreamProviderProps {
   children: React.ReactNode;
   userId: string;
   enableChat?: boolean;
   enableVideo?: boolean;
 }
 
-const StreamProvider = ({
-  children,
-  userId,
-  enableChat = true,
-  enableVideo = true,
-}: StreamProviderProps) => {
-  // Connection states - always initialize to null/false
-  // Let the connection functions handle global client detection
-  const [chatClient, setChatClient] = useState<StreamChat | null>(null);
-  const [videoClient, setVideoClient] = useState<StreamVideoClient | null>(
-    null,
-  );
-  const [chatConnected, setChatConnected] = useState(false);
-  const [videoConnected, setVideoConnected] = useState(false);
-  const [isConnecting, setIsConnecting] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  // We need BOTH a ref and state for retry count: the ref (connectionAttemptsRef)
-  // is used inside setTimeout/async closures where state would be stale, while
-  // this state variable drives re-renders so the UI shows the correct attempt count.
-  const [retryCount, setRetryCount] = useState(0);
-
-  // Use ref for connection attempts to avoid stale closures in retry logic
-  const connectionAttemptsRef = useRef(0);
-  // Guard against concurrent connectUser calls (race: connectVideo resolves first,
-  // triggers re-render + effect re-run before connectChat has set globalChatClient)
-  const isChatConnectingRef = useRef(false);
-  // Track retry timeout so we can cancel on unmount
-  const retryTimeoutRef = useRef<ReturnType<typeof setTimeout>>();
-
-  const { userDetails, isLoading } = useUserData(userId);
-
-  // Token caching with expiry tracking — use ref to avoid triggering re-renders
-  // (useState here caused getCachedToken → connectChat → connectServices to
-  //  be recreated on every token fetch, making the connectUser useEffect fire
-  //  repeatedly and producing "Consecutive calls to connectUser" warnings)
-  // Each token type has its own expiry to avoid one overwriting the other's validity window.
-  const tokenCacheRef = useRef<{
-    chatToken?: string;
-    chatExpiresAt?: number;
-    videoToken?: string;
-    videoExpiresAt?: number;
-  }>({});
-
-  const isTokenValid = useCallback((type: "chat" | "video") => {
-    const cache = tokenCacheRef.current;
-    const token = type === "chat" ? cache.chatToken : cache.videoToken;
-    const expiresAt =
-      type === "chat" ? cache.chatExpiresAt : cache.videoExpiresAt;
-    if (!token || !expiresAt) return false;
-    // Check if token expires within next 5 minutes
-    return Date.now() < expiresAt - 5 * 60 * 1000;
-  }, []);
-
-  const getCachedToken = useCallback(
-    async (type: "chat" | "video"): Promise<string> => {
-      if (isTokenValid(type)) {
-        const cache = tokenCacheRef.current;
-        return type === "chat" ? cache.chatToken! : cache.videoToken!;
-      }
-
-      // Generate new token
-      const newToken =
-        type === "chat"
-          ? await chatTokenProvider(userId)
-          : await tokenProvider(userId);
-
-      // Cache with 50-minute expiry (tokens usually last 1 hour)
-      const expiresAt = Date.now() + 50 * 60 * 1000;
-      if (type === "chat") {
-        tokenCacheRef.current.chatToken = newToken;
-        tokenCacheRef.current.chatExpiresAt = expiresAt;
-      } else {
-        tokenCacheRef.current.videoToken = newToken;
-        tokenCacheRef.current.videoExpiresAt = expiresAt;
-      }
-
-      return newToken;
-    },
-    [userId, isTokenValid],
-  );
-
-  // Exponential backoff retry logic
-  const getRetryDelay = useCallback((attempt: number) => {
-    return Math.min(1000 * Math.pow(2, attempt), 30000); // Max 30 seconds
-  }, []);
-
-  const connectChat = useCallback(async () => {
-    if (!enableChat || !userDetails || !apiKey) return;
-
-    // Check if we already have a global client for this user - adopt it
-    if (currentUserId === userDetails.id && globalChatClient) {
-      streamLogger.debug("Adopting existing chat client", {
-        userId: userDetails.id,
-      });
-      setChatClient(globalChatClient);
-      setChatConnected(true);
-      return;
-    }
-
-    // Prevent concurrent connectUser calls (e.g. connectVideo re-render race)
-    if (isChatConnectingRef.current) {
-      streamLogger.debug("Chat connection already in progress, skipping", {
-        userId: userDetails.id,
-      });
-      return;
-    }
-
-    isChatConnectingRef.current = true;
-
-    try {
-      streamLogger.debug("Connecting to Stream Chat", {
-        userId: userDetails.id,
-      });
-
-      const client = StreamChat.getInstance(apiKey);
-
-      // If the singleton is already connected to this user (e.g. StreamVideoClient
-      // connected it internally), adopt it directly without calling connectUser again.
-      if (client.userID && client.userID === userDetails.id) {
-        streamLogger.debug("Adopting already-connected Stream Chat singleton", {
-          userId: userDetails.id,
-        });
-        globalChatClient = client;
-        currentUserId = userDetails.id;
-        setChatClient(client);
-        setChatConnected(true);
-        return;
-      }
-
-      // Ensure user exists in Stream's database (only if not synced before)
-      if (!clientSyncCompletedUsers.has(userDetails.id)) {
-        try {
-          await upsertUserToStream(userDetails.id);
-          streamLogger.debug("User upserted to Stream", {
-            userId: userDetails.id,
-          });
-        } catch (upsertError) {
-          streamLogger.warn("User upsert failed, continuing", {
-            userId: userDetails.id,
-          });
-        }
-      }
-
-      const streamRole = mapRoleToStream(userDetails.role);
-
-      await client.connectUser(
-        {
-          id: userDetails.id,
-          name: userDetails.name ?? userDetails.id,
-          image: userDetails.image ?? undefined,
-          role: streamRole,
-        },
-        () => getCachedToken("chat"),
-      );
-
-      // Store in global references
-      globalChatClient = client;
-      currentUserId = userDetails.id;
-
-      setChatClient(client);
-      setChatConnected(true);
-
-      // Initial channel sync — once per user per browser session.
-      // The in-memory Set resets on page reload (client module re-evaluation),
-      // so we persist to sessionStorage to survive refreshes within the same tab.
-      const syncKey = `stream_sync_${userDetails.id}`;
-      const alreadySynced =
-        clientSyncCompletedUsers.has(userDetails.id) ||
-        (typeof sessionStorage !== "undefined" &&
-          sessionStorage.getItem(syncKey) === "1");
-
-      if (!alreadySynced) {
-        try {
-          streamLogger.info("Starting initial channel sync", {
-            userId: userDetails.id,
-          });
-          await syncUserEventChannels(userDetails.id);
-          clientSyncCompletedUsers.add(userDetails.id);
-          if (typeof sessionStorage !== "undefined") {
-            sessionStorage.setItem(syncKey, "1");
-          }
-          streamLogger.info("Initial channel sync completed", {
-            userId: userDetails.id,
-          });
-        } catch (syncError) {
-          streamLogger.warn("Channel sync failed", { userId: userDetails.id });
-          // Mark in-memory to avoid retry within same component lifecycle,
-          // but don't persist to sessionStorage so next load retries.
-          clientSyncCompletedUsers.add(userDetails.id);
-        }
-      } else {
-        streamLogger.debug("Skipping channel sync (already completed)", {
-          userId: userDetails.id,
-        });
-      }
-
-      streamLogger.info("Chat connection established", {
-        userId: userDetails.id,
-      });
-    } catch (error) {
-      streamLogger.warn("Chat connection failed (will retry)", {
-        userId: userDetails.id,
-      });
-      setChatConnected(false);
-      throw error;
-    } finally {
-      isChatConnectingRef.current = false;
-    }
-  }, [enableChat, userDetails, apiKey, getCachedToken]);
-
-  const connectVideo = useCallback(async () => {
-    if (!enableVideo || !userDetails || !apiKey) return;
-
-    // Check if we already have a global client for this user - adopt it
-    if (currentUserId === userDetails.id && globalVideoClient) {
-      streamLogger.debug("Adopting existing video client", {
-        userId: userDetails.id,
-      });
-      setVideoClient(globalVideoClient);
-      setVideoConnected(true);
-      return;
-    }
-
-    try {
-      streamLogger.debug("Connecting to Stream Video", {
-        userId: userDetails.id,
-      });
-
-      const client = new StreamVideoClient({
-        apiKey: apiKey,
-        user: {
-          id: userDetails.id,
-          name: userDetails.name ?? userDetails.id,
-          image: userDetails.image ?? undefined,
-        },
-        tokenProvider: () => getCachedToken("video"),
-      });
-
-      // Store in global reference
-      globalVideoClient = client;
-      currentUserId = userDetails.id;
-
-      setVideoClient(client);
-      setVideoConnected(true);
-      streamLogger.info("Video connection established", {
-        userId: userDetails.id,
-      });
-    } catch (error) {
-      streamLogger.error("Video connection failed", error, {
-        userId: userDetails.id,
-      });
-      setVideoConnected(false);
-      throw error;
-    }
-  }, [enableVideo, userDetails, apiKey, getCachedToken]);
-
-  // Full disconnect - only call when user changes or app unmounts
-  const disconnect = useCallback(
-    async (clearGlobal = false) => {
-      const promises = [];
-
-      if (chatClient) {
-        promises.push(
-          chatClient.disconnectUser().then(() => {
-            streamLogger.debug("Chat client disconnected");
-            setChatClient(null);
-            setChatConnected(false);
-            if (clearGlobal) {
-              globalChatClient = null;
-            }
-          }),
-        );
-      }
-
-      if (videoClient) {
-        promises.push(
-          videoClient.disconnectUser().then(() => {
-            streamLogger.debug("Video client disconnected");
-            setVideoClient(null);
-            setVideoConnected(false);
-            if (clearGlobal) {
-              globalVideoClient = null;
-            }
-          }),
-        );
-      }
-
-      await Promise.all(promises);
-      if (clearGlobal) {
-        currentUserId = null;
-        tokenCacheRef.current = {}; // Clear token cache
-        connectionAttemptsRef.current = 0; // Reset attempts
-      }
-    },
-    [chatClient, videoClient],
-  );
-
-  // Stable connectServices function using ref pattern for retry logic
-  const connectServices = useCallback(async () => {
-    if (isLoading || !userDetails) return;
-
-    setIsConnecting(true);
-    setError(null);
-
-    try {
-      const promises = [];
-      // Always try to connect - the functions will handle global client detection
-      if (enableChat) promises.push(connectChat());
-      if (enableVideo) promises.push(connectVideo());
-
-      await Promise.all(promises);
-      connectionAttemptsRef.current = 0; // Reset on success
-      setRetryCount(0);
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : "Connection failed";
-      setError(errorMessage);
-
-      // Implement exponential backoff retry using ref
-      connectionAttemptsRef.current += 1;
-      setRetryCount(connectionAttemptsRef.current); // Sync state for UI display
-      const currentAttempts = connectionAttemptsRef.current;
-
-      if (currentAttempts < 5) {
-        // Max 5 attempts
-        const delay = getRetryDelay(currentAttempts);
-        streamLogger.debug(`Retrying connection in ${delay}ms`, {
-          attempt: currentAttempts,
-        });
-        setIsConnecting(false);
-        retryTimeoutRef.current = setTimeout(() => {
-          // Re-run connection (the ref ensures we get current attempt count)
-          connectServices();
-        }, delay);
-        return;
-      } else {
-        streamLogger.error("Max connection attempts reached", error);
-      }
-    } finally {
-      setIsConnecting(false);
-    }
-  }, [
-    isLoading,
-    userDetails,
-    enableChat,
-    enableVideo,
-    connectChat,
-    connectVideo,
-    getRetryDelay,
-  ]);
-
-  const retryConnection = useCallback(() => {
-    connectionAttemptsRef.current = 0;
-    setError(null);
-    connectServices();
-  }, [connectServices]);
-
-  // Initialize connections.
-  // connectServices/disconnect are in deps and may cause re-fires when their
-  // useCallback identities change, but this is safe because:
-  // - connectChat guards with globalChatClient + currentUserId check (no-op if already connected)
-  // - syncUserEventChannels is guarded by sessionStorage (no-op after first sync)
-  useEffect(() => {
-    if (!isLoading && userDetails && apiKey) {
-      // Check if user changed - if so, disconnect old user first
-      if (currentUserId && currentUserId !== userDetails.id) {
-        streamLogger.info("User changed, reconnecting", {
-          from: currentUserId,
-          to: userDetails.id,
-        });
-        disconnect(true).then(() => {
-          connectServices();
-        });
-      } else {
-        connectServices();
-      }
-    }
-
-    // Don't disconnect on unmount - keep global clients alive for tab switching
-    // Only disconnect when user explicitly logs out or changes
-    return () => {
-      // Cancel any pending retry timeout to avoid state updates after unmount
-      if (retryTimeoutRef.current) {
-        clearTimeout(retryTimeoutRef.current);
-        retryTimeoutRef.current = undefined;
-      }
-      // Intentionally not calling disconnect() here
-      // Global clients are reused across component remounts
-    };
-  }, [userDetails?.id, isLoading, apiKey, connectServices, disconnect]);
-
-  // Connection state for context
-  const connectionState: StreamConnectionState = {
-    chatConnected,
-    videoConnected,
-    isConnecting,
-    error,
-    retryConnection,
-  };
-
-  // Loading state
-  if (
-    (enableChat && !chatClient && !error) ||
-    (enableVideo && !videoClient && !error) ||
-    isConnecting
-  ) {
-    return (
-      <div className="flex items-center justify-center min-h-[200px]">
-        <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-primary"></div>
-        {isConnecting && (
-          <p className="ml-4 text-sm text-gray-600">Connecting to Stream...</p>
-        )}
-      </div>
-    );
-  }
-
-  // Retry-in-progress state (between retries, not yet exhausted)
-  if (error && retryCount > 0 && retryCount < 5) {
-    return (
-      <div className="flex items-center justify-center min-h-[200px]">
-        <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-primary"></div>
-        <p className="ml-4 text-sm text-gray-600">
-          Retrying connection (attempt {retryCount}/5)...
-        </p>
-      </div>
-    );
-  }
-
-  // Error state
-  if (error && connectionAttemptsRef.current >= 5) {
-    return (
-      <div className="flex flex-col items-center justify-center min-h-[200px] p-4">
-        <div className="text-red-600 text-center">
-          <h3 className="font-semibold mb-2">Connection Failed</h3>
-          <p className="text-sm mb-4">{error}</p>
-          <button
-            onClick={retryConnection}
-            className="px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700"
-          >
-            Retry Connection
-          </button>
-        </div>
-      </div>
-    );
-  }
-
-  // Render providers
-  let content = children;
-
-  // Wrap with video provider if enabled and connected
-  if (enableVideo && videoClient) {
-    content = <StreamVideo client={videoClient}>{content}</StreamVideo>;
-  }
-
-  // Wrap with chat provider if enabled and connected
-  if (enableChat && chatClient) {
-    content = <Chat client={chatClient}>{content}</Chat>;
-  }
-
+// Loading strategy (deliberate, per-scenario — NOT blanket lazy):
+//
+//  • SDK code-split (lazy chunk): YES. The Stream SDK is heavy and the dashboard
+//    LANDING route is always /home, which renders no chat/video UI. Splitting the
+//    SDK + its two stylesheets into a chunk keeps them out of /home's synchronous
+//    bundle, so /home parses/paints without paying for the SDK up front.
+//
+//  • When does that chunk actually load? On dashboard routes the provider is
+//    mounted by the LAYOUT, so the chunk begins downloading on /home too — but
+//    IN PARALLEL, off the critical bundle, not blocking first paint. This is the
+//    right call (deferred/parallel, not "only when chat opens") because the
+//    consultant sidebar shows a chat-unread badge on every route incl. /home,
+//    which needs the chat client connected. Gating the whole provider behind
+//    "user opened chat" would break that badge. The actual connect (websocket)
+//    is further deferred to requestIdleCallback inside the impl (see #248 there).
+//
+//  • ssr:false: the SDK is browser-only (websockets/WebRTC); SSR-ing it is wasted
+//    server work and hydration mismatch risk.
+//
+// Children are rendered through the impl, which renders them DIRECTLY even before
+// connection completes (it no longer blocks on a spinner until connected). During
+// the brief chunk-download window the fallback below shows — this is strictly
+// shorter than the previous behavior, which blocked children until BOTH clients
+// finished their websocket handshake.
+function StreamProviderLoading() {
   return (
-    <StreamErrorBoundary
-      onError={(error, errorInfo) => {
-        streamLogger.error("Stream Provider Error", error, {
-          componentStack: errorInfo.componentStack,
-        });
-        setError(error.message);
-      }}
-      enableRetry={true}
-    >
-      <StreamConnectionContext.Provider value={connectionState}>
-        {content}
-      </StreamConnectionContext.Provider>
-    </StreamErrorBoundary>
+    <div className="flex items-center justify-center min-h-[200px]">
+      <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-primary" />
+    </div>
   );
+}
+
+const LazyStreamProviderImpl = dynamic(
+  () => import("@/providers/StreamProviderImpl"),
+  { ssr: false, loading: () => <StreamProviderLoading /> },
+);
+
+/**
+ * SDK-free shell. Renders children through the lazily-loaded implementation,
+ * which provides the chat/video contexts + connection lifecycle once its chunk
+ * loads. The impl renders children directly even before connection completes;
+ * video consumers already guard a null client, and chat consumers only render on
+ * the chat route (under <Chat>).
+ */
+const StreamProvider = (props: StreamProviderProps) => {
+  return <LazyStreamProviderImpl {...props} />;
 };
 
 export default StreamProvider;
